% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_label.R
\name{attach_var_label}
\alias{attach_var_label}
\title{Attach variable labels}
\usage{
attach_var_label(df)
}
\arguments{
\item{df}{An extracted data frame from the scorecard merged files}
}
\value{
A data frame, with proper variable labels
}
\description{
Attach variable labels
}
\examples{
\dontrun{
attach_var_label(scorecard::mf2014_15)
}
}
