% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{College Scorecard Data codebook}
\format{A data frame}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
codebook
}
\description{
\code{codebook} contains infomation (variable names, labels, label values)
for variables from College Scorecard Data
}
\keyword{datasets}

