% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cohort_map}
\alias{cohort_map}
\title{Cohort Map}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43494 rows and 3 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
cohort_map
}
\description{
Provides descriptions of which years of data are described for each variable
in the merged files.
}
\keyword{datasets}
