% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_derive.R
\name{drvar_behavior}
\alias{drvar_behavior}
\title{variables derived from behavior data sets}
\usage{
drvar_behavior(dat)
}
\value{
derive varibles
}
\description{
This function creates behavior derive variables from consumption and media date sets
}
