% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_derive.R
\name{drvar_k3}
\alias{drvar_k3}
\title{variables derived from 3keys}
\usage{
drvar_k3(dat)
}
\value{
derive varibles
}
\description{
This function creates derive variables from 3key variables (id, cell, mail and user_date)
}
