% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_ent_indx_gini.R
\name{ig}
\alias{ig}
\title{Index Gini}
\usage{
ig(dt, y, x = NULL, order = TRUE)
}
\arguments{
\item{dt}{A data frame with both x (predictor/feature) and y (response/label) variables.}

\item{y}{Name of y variable.}

\item{x}{Name of x variables. Default is NULL. If x is NULL, then all variables except y are counted as x variables.}

\item{order}{Logical, default is TRUE. If it is TRUE, the output will descending order via gini}
}
\value{
gini index
}
\description{
This function calculates gini index (in CART) for multiple x variables.
}
\examples{
# Load German credit data
data(germancredit)

# Information Entropy
dt_gini = ig(germancredit, y = "creditability")

}
