% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged1997}
\alias{merged1997}
\title{College Scorecard Data, 1997}
\format{A data frame}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
merged1997
}
\description{
\code{merged1997} contains all the variables (as of Sep. 17, 2015)
from College Scorecard Data, 1997
}
\keyword{datasets}

