% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged1998}
\alias{merged1998}
\title{College Scorecard Data, 1998}
\format{A data frame}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
merged1998
}
\description{
\code{merged1998} contains all the variables (as of Sep. 17, 2015)
from College Scorecard Data, 1998
}
\keyword{datasets}

