% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged1999}
\alias{merged1999}
\title{College Scorecard Data, 1999}
\format{A data frame}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
merged1999
}
\description{
\code{merged1999} contains all the variables (as of Sep. 17, 2015)
from College Scorecard Data, 1999
}
\keyword{datasets}

