% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged2004}
\alias{merged2004}
\title{College Scorecard Data, 2004}
\format{A data frame}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
merged2004
}
\description{
\code{merged2004} contains all the variables (as of Sep. 17, 2015)
from College Scorecard Data, 2004
}
\keyword{datasets}

