% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merged2009}
\alias{merged2009}
\title{College Scorecard Data, 2009}
\format{A data frame}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
merged2009
}
\description{
\code{merged2009} contains all the variables (as of Sep. 17, 2015)
from College Scorecard Data, 2009
}
\keyword{datasets}

