% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf1996_97}
\alias{mf1996_97}
\title{College Scorecard Data, 1996-97}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6794 rows and 1744 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf1996_97
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
