% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf1997_98}
\alias{mf1997_98}
\title{College Scorecard Data, 1997-98}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6699 rows and 381 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf1997_98
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
