% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf1998_99}
\alias{mf1998_99}
\title{College Scorecard Data, 1998-99}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6480 rows and 580 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf1998_99
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
