% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf1999_00}
\alias{mf1999_00}
\title{College Scorecard Data, 1999-00}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6466 rows and 806 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf1999_00
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
