% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2000_01}
\alias{mf2000_01}
\title{College Scorecard Data, 2000-01}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6478 rows and 988 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2000_01
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
