% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2002_03}
\alias{mf2002_03}
\title{College Scorecard Data, 2002-03}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6546 rows and 1191 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2002_03
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
