% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2003_04}
\alias{mf2003_04}
\title{College Scorecard Data, 2003-04}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6585 rows and 1218 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2003_04
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
