% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2004_05}
\alias{mf2004_05}
\title{College Scorecard Data, 2004-05}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6660 rows and 1415 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2004_05
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
