% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2005_06}
\alias{mf2005_06}
\title{College Scorecard Data, 2005-06}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6824 rows and 1440 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2005_06
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
