% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2006_07}
\alias{mf2006_07}
\title{College Scorecard Data, 2006-07}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6848 rows and 1744 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2006_07
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
