% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2007_08}
\alias{mf2007_08}
\title{College Scorecard Data, 2007-08}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6890 rows and 1472 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2007_08
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
