% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2009_10}
\alias{mf2009_10}
\title{College Scorecard Data, 2009-10}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7149 rows and 1744 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2009_10
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
