% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2010_11}
\alias{mf2010_11}
\title{College Scorecard Data, 2010-11}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7414 rows and 1580 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2010_11
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
