% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2012_13}
\alias{mf2012_13}
\title{College Scorecard Data, 2012-13}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7793 rows and 1603 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2012_13
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
