% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2013_14}
\alias{mf2013_14}
\title{College Scorecard Data, 2013-14}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7804 rows and 1743 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2013_14
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
