% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mf2015_16}
\alias{mf2015_16}
\title{College Scorecard Data, 2015-16}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7593 rows and 1629 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
mf2015_16
}
\description{
Contains all the variables from College Scorecard Merged Files.
}
\keyword{datasets}
