% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perf_plot}
\alias{perf_plot}
\title{ks value and ks curve}
\usage{
perf_plot(label, pred, groupnum = 20, type = c("ks"), plot = TRUE,
  seed = 186)
}
\arguments{
\item{label}{label values, such as 0s and 1s.}

\item{pred}{predicted probability values.}

\item{groupnum}{the number of group numbers, default: 20.}

\item{type}{performance plot types: "ks","lift","roc","pr". Default: c("ks").}

\item{plot}{logical value, default: TRUE.}

\item{seed}{seed value for random sort data frame, defalut: 186.}
}
\value{
ks value and ks curve.
}
\description{
This function calculate kolmogorov-smirnow(ks) value and plot ks curve based on provided label and predicted probability values.
}
\examples{
perf_plot(label, pred)

}
