% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perf_score_bad}
\alias{perf_score_bad}
\title{Score distribution and bad probability}
\usage{
perf_score_bad(label, score, title = "Socre distribution and bad probability",
  positive = "bad|1", x_limits = c(100, 700), x_tick_break = 50,
  seed = 186)
}
\arguments{
\item{label}{label values, such as 0s and 1s.}

\item{score}{credit score.}

\item{title}{plot title, default "Socre distribution and bad probability".}

\item{positive}{Value of positive class, default "bad|1".}

\item{x_limits}{x-axis limits, default c(0, 800)}

\item{x_tick_break}{xaxis ticker break, default 100}

\item{seed}{An integer. The specify seed is used for random sorting data, default: 186.}
}
\description{
\code{perf_score_bad} provides a plot with score distribution and bad probability trends.
}
