% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecards.R
\name{scorecards}
\alias{scorecards}
\title{credit score & scorecards}
\usage{
scorecards(dt_woe, y, bins, model, p0 = 600, odds0 = 1/60, pdo = 50)
}
\arguments{
\item{dt_woe}{input data that converted into woe}

\item{y}{name of y variable.}

\item{bins}{Binning information generated from woebin function}

\item{model}{fitted glm model}

\item{p0}{target points, default 600}

\item{odds0}{target odds, default 1/60}

\item{pdo}{Points to Double the Odds, default 50}
}
\value{
score and scorecards
}
\description{
This function calculates credit score based on predited probability.
}
\examples{
# # Traditional Credit Scoring Using Logistic Regression
# library(woebin)
# # load germancredit data
# data("germancredit")
#
# # rename creditability as y
# dt <- data.table(germancredit)[, `:=`(
#   y = ifelse(creditability == "bad", 1, 0),
#   creditability = NULL
# )]
#
# # woe binning ------
# bins <- woebin(dt, "y")$bins
# dt_woe <- woebin_ply(dt, "y", bins)
#
# # glm ------
# m1 <- glm( y ~ ., family = "binomial", data = dt_woe)
# # summary(m1)
#
# # Select a formula-based model by AIC
# m_step <- step(m1, direction="both")
# m2 <- eval(m_step$call)
# # summary(m2)
#
# # performance ------
# # predicted proability
# dt_woe$pred <- predict(m2, type='response', dt_woe)
#
# # performace
# # ks & roc plot
# perf_plot(dt_woe$y, dt_woe$pred)
#
# # score
# dt_woe$score <- scorecards(dt_woe, "y", bins, m2)$score
#
# # cards
# cards <- scorecards(dt_woe, "y", bins, m2)$cards

}
