% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_eda.R
\name{var_describe}
\alias{var_describe}
\title{Variable Describe}
\usage{
var_describe(dt)
}
\arguments{
\item{dt}{A data frame.}
}
\description{
This function provides descriptive statistic for exploratory data analysis.
}
\examples{
data("germancredit")

eda = var_describe(germancredit)

}
