% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_filter.R
\name{var_filter}
\alias{var_filter}
\title{Variable Filter}
\usage{
var_filter(dt, y, x = NULL, lims = list(missing_rate = 0.95, identical_rate
  = 0.95, info_value = 0.02), var_rm = NULL, var_kp = NULL,
  var_rm_reason = FALSE, positive = "bad|1", ...)
}
\arguments{
\item{dt}{A data frame with both x (predictor/feature) and y (response/label) variables.}

\item{y}{Name of y variable.}

\item{x}{Name of x variables. Defaults to NULL. If x is NULL, then all columns except y are counted as x variables.}

\item{lims}{A list of variable filters' thresholds.
\itemize{
\item \code{missing_rate} The missing rate of kept variables should <= 0.95 by defaults.
\item \code{identical_rate} The identical value rate (excluding NAs) of kept variables should <= 0.95 by defaults.
\item \code{info_value} The information value (iv) of kept variables should >= 0.02 by defaults.
}}

\item{var_rm}{Name of force removed variables, Defaults to NULL.}

\item{var_kp}{Name of force kept variables, Defaults to NULL.}

\item{var_rm_reason}{Logical, Defaults to FALSE.}

\item{positive}{Value of positive class, Defaults to "bad|1".}

\item{...}{Additional parameters.}
}
\value{
A data frame with columns for y and selected x variables, and a data frame with columns for remove reason if var_rm_reason is TRUE.
}
\description{
This function filter variables base on specified conditions, such as missing rate, identical value rate, information value.
}
\examples{
# Load German credit data
data(germancredit)

# variable filter
dt_sel = var_filter(germancredit, y = "creditability")
dim(dt_sel)

# return the reason of varaible removed
dt_sel2 = var_filter(germancredit, y = "creditability", var_rm_reason = TRUE)
lapply(dt_sel2, dim)

str(dt_sel2$dt)
str(dt_sel2$rm)

# keep columns manually, such as rowid
germancredit$rowid = row.names(germancredit)
dt_sel3 = var_filter(germancredit, y = "creditability", var_kp = 'rowid')

# remove columns manually
dt_sel4 = var_filter(germancredit, y = "creditability", var_rm = 'rowid')

}
