% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woebin.R
\name{woebin_plot}
\alias{woebin_plot}
\title{binning visualization}
\usage{
woebin_plot(bins)
}
\arguments{
\item{bins}{binning generated via \code{\line{woebin}}}
}
\description{
This function visualizes the binning results generated via \code{\line{woebin}}
}
\examples{
data(germancredit)
bins <- woebin(germancredit, y="creditability")$bins

plotlist <- woebin_plot(bins)
plotlist

}
