% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{College Scorecard Data codebook}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1975 rows and 9 columns.}
\source{
\url{https://collegescorecard.ed.gov/data}
}
\usage{
codebook
}
\description{
Contains variable names, labels, label values for the College Scorecard Data.
}
\keyword{datasets}
