% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Variable Describe}
\usage{
describe(dt)
}
\arguments{
\item{dt}{A data frame.}
}
\description{
This function provides descriptive statistic for exploratory data analysis.
}
\examples{
library(data.table)

data("germancredit")
dat = rbind(
  setDT(germancredit),
  data.table(creditability=sample(c("good","bad"),100,replace=TRUE)),
  fill=TRUE)

eda = describe(dat)
eda

}
