% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.R
\name{iv}
\alias{iv}
\title{Information Value}
\usage{
iv(dt, y, x = NA, positive = "bad|1", order = "TRUE")
}
\arguments{
\item{dt}{A data frame with both x (predictor/feature) and y (response/label) variables.}

\item{y}{Name of y variable.}

\item{x}{Name vector of x variables. Default NA. If x is NA, all variables exclude y will counted as x variables.}

\item{positive}{Value of positive class, default "bad|1".}

\item{order}{Logical. If it is TRUE, return descending sorted iv values.}
}
\value{
IV of all x variables.
}
\description{
This function calculates information value (IV) for all x variables.
}
\examples{
# Load German credit data
data(germancredit)

# information values
iv(germancredit, y = "creditability")

}
