% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.R
\name{iv}
\alias{iv}
\title{information values}
\usage{
iv(dt, y, x = NA, positive = "bad|1", order = "TRUE")
}
\arguments{
\item{dt}{Name of data.frame/data.table with input data.}

\item{y}{Name of y variable.}

\item{x}{Name vector of x variables.}

\item{positive}{The positive/bad target event, such as "bad" or 1.}
}
\value{
List of woe and iv data tables.
}
\description{
This function calculates IV.
}
\examples{
# Load German credit data
data(germancredit)

# information values
iv(germancredit, y = "creditability")
}
