% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.R
\name{iv}
\alias{iv}
\title{information values (iv)}
\usage{
iv(dt, y, x, positive)
}
\arguments{
\item{dt}{Name of data.frame/data.table with input data.}

\item{y}{Name of y variable.}

\item{x}{Name vector of x variables.}

\item{positive}{The positive/bad target event, such as "bad" or 1.}
}
\value{
List of woe and iv data tables.
}
\description{
This function calculate woe and iv values.
}
\examples{
# Load German credit data and create good and bad series
data(germancredit)
dt <- germancredit[, c('creditability', 'credit.amount', 'age.in.years')]

# iv(dt, y)
iv(dt, y = "creditability")
}
