% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{one_hot}
\alias{one_hot}
\title{One Hot Encoding}
\usage{
one_hot(dt, cols_skip = NULL, cols_encode = NULL,
  factor_to_integer = FALSE, nacol_rm = TRUE, replace_na = NULL)
}
\arguments{
\item{dt}{A data frame.}

\item{cols_skip}{Name of categorical variables that will skip and without doing one-hot encoding. Default is NULL.}

\item{cols_encode}{Name of categorical variables to be one-hot encoded, default is NULL. If it is NULL, then all categorical variables except in cols_skip are counted.}

\item{factor_to_integer}{Logical. Converting factor variables to integer. Default is FALSE.}

\item{nacol_rm}{Logical. One-hot encoding on categorical variable contains missing values, whether to remove the column generated to indicate the presence of NAs. Default is TRUE.}

\item{replace_na}{Replace missing values with the mean/median value of the variable in which they occur, or with specified value such as -1. Default is NULL.}
}
\description{
One-hot encoding on categorical variables. It is not needed when creating a standard scorecard model, but required in models that without doing woe transformation.
}
\examples{
# load germancredit data
data(germancredit)

library(data.table)
dat = rbind(
  germancredit,
  data.table(creditability=sample(c("good","bad"),10,replace=TRUE)),
  fill=TRUE)

dat_onehot = one_hot(dat, cols_skip = 'creditability')
dat_onehot2 = one_hot(dat, cols_skip = 'creditability', replace_na = -1)
dat_onehot3 = one_hot(dat, cols_skip = 'creditability', replace_na = 'mean')

}
