% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perf_plot}
\alias{perf_plot}
\title{renamed as perf_eva}
\usage{
perf_plot(label, pred, title = "train", groupnum = 20, type = c("ks",
  "roc"), show_plot = TRUE, seed = 186)
}
\arguments{
\item{label}{Label values, such as 0s and 1s, 0 represent for good and 1 for bad.}

\item{pred}{Predicted probability values.}

\item{title}{Title of plot, default "train".}

\item{groupnum}{The group numbers when calculating bad probability, default 20.}

\item{type}{Types of performance plot, such as "ks", "lift", "roc", "pr". Default c("ks", "roc").}

\item{show_plot}{Logical value, default TRUE. It means whether to show plot.}

\item{seed}{An integer. The specify seed is used for random sorting data, default: 186.}
}
\description{
The function perf_plot has renamed as perf_eva.
}
