% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_scorecard.R
\name{perf_psi}
\alias{perf_psi}
\title{psi}
\usage{
perf_psi(train_labelscore, test_labelscore, y, title = "PSI", groupnum = 20,
  positive = "bad|1", plot = TRUE, plot_total = FALSE, seed = 186)
}
\arguments{
\item{train_labelscore}{train data with only label and score}

\item{test_labelscore}{test data with only label and score}

\item{y}{Name of y variable.}

\item{title}{plot title, default "train".}

\item{groupnum}{the number of group numbers, default: 20.}

\item{positive}{Name of positive class, defaults: bad or 1.}

\item{plot}{logical value, default TRUE. It means whether to display plot.}

\item{plot_total}{logical value, default FALSE, which means not display the line of total PSI}

\item{seed}{seed value for random sort data frame, defalut: 186.}
}
\value{
psi
}
\description{
This function provides population stability index (PSI).
}
