% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{scorecard}
\alias{scorecard}
\title{create scorecard}
\usage{
scorecard(bins, model, p0 = 600, odds0 = 1/60, pdo = 50)
}
\arguments{
\item{bins}{Binning information generated from \code{woebin} function.}

\item{model}{A glm model object.}

\item{p0}{Target points, default 600.}

\item{odds0}{Target odds, default 1/60.}

\item{pdo}{Points to Double the Odds, default 50.}
}
\value{
scorecard
}
\description{
This function creates scorecard based on the results from \code{woebin} and \code{glm}.
}
\examples{
# # Traditional Credit Scoring Using Logistic Regression
# # load germancredit data
# data("germancredit")
#
# # rename creditability as y
# dt <- setDT(germancredit)[, `:=`(
#   y = ifelse(creditability == "bad", 1, 0),
#   creditability = NULL
# )]
#
# # woe binning ------
# bins <- woebin(dt, "y")
# dt_woe <- woebin_ply(dt, bins)
#
# # glm ------
# m1 <- glm( y ~ ., family = "binomial", data = dt_woe)
# # summary(m1)
#
# # Select a formula-based model by AIC
# m_step <- step(m1, direction="both")
# m2 <- eval(m_step$call)
# # summary(m2)
#
# # performance ------
# # predicted proability
# dt_woe$pred <- predict(m2, type='response', dt_woe)
#
# # performace
# # ks & roc plot
# perf_plot(dt_woe$y, dt_woe$pred)
#
# # card
# card <- scorecard(bins, m2)
#
# # score
# dt_woe$score <- scorecard_ply(dt, card)

}
