% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{scorecard_ply}
\alias{scorecard_ply}
\title{calculates credit score}
\usage{
scorecard_ply(dt, card, only_total_score = TRUE)
}
\arguments{
\item{dt}{Original data}

\item{card}{Scorecard generated from \code{scorecard}.}

\item{only_total_score}{Logical, default TRUE. If it is TRUE, return total credit score only; if FALSE, return both total credit score and score points of each variables.}
}
\value{
credit score points
}
\description{
This function calculates credit score using the scorecard.
}
\examples{
# # Traditional Credit Scoring Using Logistic Regression
# # load germancredit data
# data("germancredit")
#
# # rename creditability as y
# dt <- setDT(germancredit)[, `:=`(
#   y = ifelse(creditability == "bad", 1, 0),
#   creditability = NULL
# )]
#
# # woe binning ------
# bins <- woebin(dt, "y")
# dt_woe <- woebin_ply(dt, bins)
#
# # glm ------
# m1 <- glm( y ~ ., family = "binomial", data = dt_woe)
# # summary(m1)
#
# # Select a formula-based model by AIC
# m_step <- step(m1, direction="both")
# m2 <- eval(m_step$call)
# # summary(m2)
#
# # performance ------
# # predicted proability
# dt_woe$pred <- predict(m2, type='response', dt_woe)
#
# # performace
# # ks & roc plot
# perf_plot(dt_woe$y, dt_woe$pred)
#
# # card
# card <- scorecard(bins, m2)
#
# # score
# dt_woe$score <- scorecard_ply(dt, card)

}
