% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_scorecard.R
\name{scorecards}
\alias{scorecards}
\title{credit score & scorecards}
\usage{
scorecards(dt_woe, y, bins, model, p0 = 600, odds0 = 1/60, pdo = 50)
}
\arguments{
\item{dt_woe}{input data that converted into woe}

\item{y}{name of y variable.}

\item{bins}{Binning information generated from woebin function}

\item{model}{fitted glm model}

\item{p0}{target points, default 600}

\item{odds0}{target odds, default 1/60}

\item{pdo}{Points to Double the Odds, default 50}
}
\value{
score and scorecards
}
\description{
This function calculates credit score based on predited probability.
}
