% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitdf.R
\name{split_df}
\alias{split_df}
\title{Split a dataset}
\usage{
split_df(dt, y = NULL, ratio = 0.7, seed = 186)
}
\arguments{
\item{dt}{A data frame.}

\item{y}{Name of y variable, default is NULL. The input data will split based on the predictor y, if it is provide.}

\item{ratio}{A numeric value, default is 0.7. It indicates the ratio of total rows contained in one split, must less than 1.}

\item{seed}{A random seed, default is 186.}
}
\description{
Split a dataset
}
\examples{
# Load German credit data
data(germancredit)

dt_list = split_df(germancredit, y="creditability")
train = dt_list$train
test = dt_list$test

}
