% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_var_filter.R
\name{var_filter}
\alias{var_filter}
\title{variable filter}
\usage{
var_filter(dt, y, x = "", iv_limit = 0.02, na_perc_limit = 0.95,
  uniqueone_rm = TRUE)
}
\arguments{
\item{dt}{Name of input data}

\item{y}{Name of y variable.}

\item{x}{Name vector of x variables, default "".}

\item{iv_limit}{The minimun IV of variables that are kept, default 0.02.}

\item{na_perc_limit}{The maximum NA percent of variables that are kept, default 0.95.}

\item{uniqueone_rm}{Logical, default TRUE, it remove variables that have only one unique elements (including NAs).}
}
\description{
This function filter variables base on their iv, na percentage and length of unique elements.
}
\examples{
# Load German credit data and create good and bad series
data(germancredit)
dt <- germancredit[, c('creditability', 'credit.amount', 'age.in.years', 'present.employment.since')]

# variable filter
var_filter(dt, y = "creditability")

}
