% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woebin.R
\name{woebin}
\alias{woebin}
\title{woe binning}
\usage{
woebin(dt, y, x = "", binnum = 6, method = "tree",
  min_perc_total = 0.01, positive = "bad|1", print_step = FALSE)
}
\arguments{
\item{dt}{Name of input data}

\item{y}{Name of y variable.}

\item{x}{Name vector of x variables, defaults: "".}

\item{binnum}{Number of binning, defaults: 6.}

\item{min_perc_total}{The share of initial binning class on total number.}

\item{positive}{Name of positive class, defaults: bad or 1.}
}
\value{
List of binnig for each variable.
}
\description{
This function generates a list of binnings based on tree-like segmentation for both numeric and character variables.
}
\examples{
# Load German credit data and create good and bad series
data(germancredit)
dt <- germancredit[, c('creditability', 'credit.amount', 'age.in.years', 'present.employment.since')]

woebin(dt, y = "creditability")
}
