% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woebin_adj.R
\name{woebin_adj}
\alias{woebin_adj}
\title{binning adjustment}
\usage{
woebin_adj(dt, y, x = "")
}
\arguments{
\item{dt}{name of input data frame.}

\item{y}{name of y variable.}

\item{x}{names of x variables, default: all column names except y.}
}
\description{
This function provides a shiny interface for binning adjustment.
}
\examples{
data(germancredit)
woebin_adj(germancredit, y="creditability")

}
