% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woebin.R
\name{woebin_adj}
\alias{woebin_adj}
\title{WOE Binning Adjustment}
\usage{
woebin_adj(bins, dt, y)
}
\arguments{
\item{bins}{A list or data frame. Binning information generated by \code{woebin}.}

\item{dt}{A data frame.}

\item{y}{Name of y variable.}
}
\description{
\code{woebin_adj} adjust the binning breaks from \code{woebin}.
}
\examples{
\dontrun{
# Load German credit data
data(germancredit)

dt1 = germancredit[, c("creditability", "credit.amount")]

bins1 = woebin(dt1, y="creditability")
break_adj = woebin_adj(bins1, dt1, "creditability")
}

}
\seealso{
\code{\link{woebin}}, \code{\link{woebin_ply}}, \code{\link{woebin_plot}}
}
