% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woebin.R
\name{woebin_adj}
\alias{woebin_adj}
\title{WOE Binning Adjustment}
\usage{
woebin_adj(bins, dt, y)
}
\arguments{
\item{bins}{A list or data frame. Binning information generated by \code{woebin}.}

\item{dt}{A data frame.}

\item{y}{Name of y variable.}
}
\description{
\code{woebin_adj} interactively adjust the binning breaks.
}
\examples{
\dontrun{
# Load German credit data
data(germancredit)

# Example II
dt1 = germancredit[, c("creditability", "credit.amount")]
bins1 = woebin(dt1, y="creditability")
breaks_adj1 = woebin_adj(bins1, dt1, y="creditability")
bins1_final = woebin(dt1, y="creditability",
                    breaks_list=breaks_adj1)

# Example II
bins = woebin(germancredit, y="creditability")
breaks_adj = woebin_adj(bins, germancredit, "creditability")
bins_final = woebin(germancredit, y="creditability",
                    breaks_list=breaks_adj)

}

}
\seealso{
\code{\link{woebin}}, \code{\link{woebin_ply}}, \code{\link{woebin_plot}}
}
