% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woebin.R
\name{woebin_plot}
\alias{woebin_plot}
\title{binning visualization}
\usage{
woebin_plot(bins, x = NULL)
}
\arguments{
\item{bins}{Binning information generated by \code{woebin}.}

\item{x}{Name vector of x variables. Default NA.}
}
\value{
A list of ggplot objects.
}
\description{
This function generates plots for count and bad probability in each bin, which are the binning information from \code{woebin}.
}
\examples{
data(germancredit)
bins <- woebin(germancredit, y="creditability")

plotlist <- woebin_plot(bins)
plotlist

}
