% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woebin.R
\name{woebin_plot}
\alias{woebin_plot}
\title{binning visualization}
\usage{
woebin_plot(bins, x = NULL, title = "")
}
\arguments{
\item{bins}{Binning information generated by \code{woebin}.}

\item{x}{Name vector of x variables. Default NA.}

\item{title}{String added to the front of plot title, default "".}
}
\value{
A list of ggplot objects.
}
\description{
\code{woebin_plot} create plots of count distribution and bad probability for each bin. The binning informations are generates by  \code{woebin}.
}
\examples{
# Load German credit data
data(germancredit)

dt <- germancredit[, c("creditability", "credit.amount")]

# binning
bins1 <- woebin(dt, y="creditability")

# binning plot
woebin_plot(bins1)

\dontrun{
bins <- woebin(germancredit, y="creditability")
plotlist <- woebin_plot(bins)
}

}
\seealso{
\code{\link{woebin}} \code{\link{woebin_ply}}
}
