% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woebin.R
\name{woebin_ply}
\alias{woebin_ply}
\title{woe binning apply}
\usage{
woebin_ply(dt, bins)
}
\arguments{
\item{dt}{A data frame.}

\item{bins}{Binning information generated by \code{woebin}.}
}
\value{
WOE value for each variable.
}
\description{
\code{woebin_ply} converts original input data to woe values based on the binning information generated by \code{woebin}.
}
\examples{
# load germancredit data
data(germancredit)

dt <- germancredit[, c("creditability", "age.in.years",
      "credit.amount", "housing", "purpose")]

# binning
bins <- woebin(dt, y = "creditability")

dt_woe <- woebin_ply(dt, bins=bins)

}
