% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woebin.R
\name{woebin_ply}
\alias{woebin_ply}
\title{binning apply}
\usage{
woebin_ply(dt, bins, y)
}
\arguments{
\item{dt}{Name of input data}

\item{bins}{Binning information generated from woebin function}

\item{y}{Name of y variable}
}
\value{
List of binnig for each variable.
}
\description{
This function applies binning via woebin to dataframe
}
\examples{

# load germancredit data
data(germancredit)
dt <- germancredit[, c('creditability', 'age.in.years', 'present.employment.since')]

# woe binning
wb <- woebin(dt, "creditability")
# woe binning apply
dt_woe <- woebin_ply(dt, wb$bins, "creditability")
}
